Fix bug where clicking outside #app would open reference
Because the #reference overlay was positioned behind the #app container,
clicking virtually anywhere outside of the #app container would trigger
a click on the #reference container, and since the #reference container
was also a toggle control for itself, the reference container would
open. However, the "App Reference" link in the top corner of the page
should be the only means of opening the reference sidebar. Therefore,
this commit replaces the bidirectional toggle classes with
unidirectional open/close classes, so that the #reference container can
only ever close itself when clicked, and never open itself.