Initialize variable values to false instead of true
This change does does affect the end behavior of the mapPermutations
function; it is only for the purpose of making the code more intuitive.
Previously, all values in the currentVarValues map were initialized to
true because they would be inverted on the first permutation (making the
first permutation of variable values all false). However, doing this
creates an unintuitive disconnect between the implementation and the
actual behavior, because the initialization values are not equal to the
values of the first permutation. Therefore, change the implementation to
initialize the values to true (which requires skipping the value
inversion for the very first permutation).