Fix adding/removing of variables/expressions
This commit fixes a bug where variables and expressions could not be
added/removed. This bug was caused by the removal of the underscore
partials in 0d46267, where the partials themselves were removed without
updating other parts of the component code accordingly. In effect, the
controller and app model could not be accessed within event handlers
because they were not being passed properly.